int to double

+++ src/BankAccount.java
@@ -2,7 +2,7 @@ public class BankAccount {
     private String name;
     private double balance;
 
-    public BankAccount(String name, int balance) {
+    public BankAccount(String name, double balance) {
         this.name = name;
         this.balance = balance;
     }
